# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Adversary backend API — scenarios, geo, runs (WebSocket ops+map), manage, generate/route, AUTH_MODE=off|session. Live channels: GET /v1/runs/{id}/ws/ops and /ws/map (Upgrade). Prometheus scrape: GET /metrics (not listed as OpenAPI paths). Architecture: docs/api.md.",
        "title": "Adversary API",
        "contact": {},
        "version": "0.9.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/runs": {
            "get": {
                "description": "Default: status=running plus recently completed/stopped (updated within 24h). Each row includes scenarioName. Use ?active=true for running-only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "List runs (Active scenarios)",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "If true, only status=running",
                        "name": "active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.RunSummary"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Starts a ready scenario. startAt shifts the schedule (schedule_offset_ms); omit for now. AUTH_MODE=session requires ownership of the scenario.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Start a run",
                "parameters": [
                    {
                        "description": "scenarioId + optional startAt",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.createRunRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.createRunResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/runs/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Get run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.RunSummary"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/runs/{id}/snapshot": {
            "get": {
                "description": "Cold-load: all events due ≤ now plus current target states (not catch-up wire shape)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Run snapshot (full due-set)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_engine.Snapshot"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/runs/{id}/stop": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Stop a run",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.RunSummary"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
        "/v1/runs/{id}/viewport": {
            "get": {
                "description": "Cold-load targets/trails in PostGIS envelope bbox ∪ includeTargetIds. zoom is advisory (LOD reserved). Eviction: client drops overlay when unwatched and outside bbox; server only sends matching targets.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "runs"
                ],
                "summary": "Map viewport snapshot",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "BBox west (WGS84)",
                        "name": "west",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "BBox south",
                        "name": "south",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "BBox east",
                        "name": "east",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "BBox north",
                        "name": "north",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "Map zoom (advisory)",
                        "name": "zoom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated off-screen tracked target ids",
                        "name": "includeTargetIds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_engine.ViewportSnapshot"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
...
        "/v1/runs/{id}/ws/ops": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "Live ops: event.ingested, alert.raised, target.updated, run.completed/stopped; catch-up = catchup.target.updated (latest target set). Redis pub/sub for multi-instance. AUTH_MODE=session requires cookie.",
                "tags": [
                    "runs"
                ],
                "summary": "Ops WebSocket channel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Run ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.errorBody"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_shammalie_adversary_apps_api_internal_engine.Snapshot": {
            "type": "object",
            "properties": {
                "asOf": {
                    "type": "string"
                },
                "criticalAlertIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "ingestedEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_scenario.SimulationEvent"
                    }
                },
                "processedEventIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "run": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.RunSummary"
                },
                "status": {
                    "type": "string"
                },
                "targetStates": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_simulation.RuntimeTargetState"
                    }
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_engine.ViewportSnapshot": {
            "type": "object",
            "properties": {
                "asOf": {
                    "type": "string"
                },
                "bbox": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_viewport.BBox"
                },
                "includeTargetIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "run": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_store.RunSummary"
                },
                "status": {
                    "type": "string"
                },
                "targetStates": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_simulation.RuntimeTargetState"
                    }
                },
                "zoom": {
                    "type": "number"
                }
            }
        },
...
        "github_com_shammalie_adversary_apps_api_internal_scenario.PositionPayload": {
            "type": "object",
            "properties": {
                "altitude": {
                    "type": "number"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "speed": {
                    "type": "number"
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_scenario.SimulationEvent": {
            "type": "object",
            "properties": {
                "at": {
                    "type": "string"
                },
                "firesAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_scenario.PositionPayload"
                },
                "targetId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_shammalie_adversary_apps_api_internal_simulation.PositionSnapshot": {
            "type": "object",
            "properties": {
                "altitude": {
                    "type": "number"
                },
                "at": {
                    "type": "string"
                },
                "course": {
                    "type": "number"
                },
                "heading": {
                    "type": "number"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "speed": {
                    "type": "number"
                }
            }
        },
        "github_com_shammalie_adversary_apps_api_internal_simulation.RuntimeTargetState": {
            "type": "object",
            "properties": {
                "appeared": {
                    "type": "boolean"
                },
                "callsign": {
                    "type": "string"
                },
                "color": {
                    "type": "string"
                },
                "lastEventAt": {
                    "type": "string"
                },
                "position": {
                    "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_simulation.PositionSnapshot"
                },
                "profile": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "revealed": {
                    "type": "boolean"
                },
                "targetId": {
                    "type": "string"
                },
                "trail": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_shammalie_adversary_apps_api_internal_simulation.PositionSnapshot"
                    }
                }
            }
        },
...
        "github_com_shammalie_adversary_apps_api_internal_store.RunSummary": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scenarioId": {
                    "type": "string"
                },
                "scenarioName": {
                    "type": "string"
                },
                "scheduleOffsetMs": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "stoppedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_shammalie_adversary_apps_api_internal_viewport.BBox": {
            "type": "object",
            "properties": {
                "east": {
                    "type": "number"
                },
                "north": {
                    "type": "number"
                },
                "south": {
                    "type": "number"
                },
                "west": {
                    "type": "number"
                }
            }
        },
...
        "internal_handler.createRunRequest": {
            "type": "object",
            "properties": {
                "scenarioId": {
                    "type": "string"
                },
                "startAt": {
                    "type": "string"
                }
            }
        },
        "internal_handler.createRunResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "runId": {
                    "type": "string"
                },
                "scenarioId": {
                    "type": "string"
                },
                "scenarioName": {
                    "type": "string"
                },
                "scheduleOffsetMs": {
                    "type": "integer"
                },
                "startAt": {
                    "type": "string"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "stoppedAt": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler.errorBody": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "adversary_session",
            "in": "cookie"
        }
    }
}